refactor: move Navigation out of Header
