+++ src/components/header.tsx
@@ -6,21 +6,8 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { cn } from '@/lib/utils';
 import { Button } from '@/components/ui/button';
-import { Icon } from '@/components/icon';
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu';
 import { Sheet, SheetContent, SheetTrigger } from '@/components/ui/sheet';
-
-const navLinks = [
-  { name: 'Rules', path: '/rules' },
-  { name: 'Booking', path: '/booking' },
-];
+import Navigation from '@/components/navigation';
 
 export default function Header() {
   const pathname = usePathname();
@@ -28,7 +15,7 @@ export default function Header() {
 
   return (
     <header className="container sticky top-0 z-10 flex h-16 items-center gap-4 border-b bg-background">
-      <nav className="hidden flex-col gap-6 text-lg font-medium md:flex md:flex-row md:items-center md:gap-5 md:text-sm lg:gap-6">
+      <div className="hidden flex-col gap-6 md:flex md:flex-row md:items-center md:gap-5 lg:gap-6">
         <Link
           href="/"
           className="flex h-8 w-8 items-center gap-2 text-lg font-semibold md:text-base"
@@ -42,21 +29,8 @@ export default function Header() {
           />
           <span className="sr-only">TTUMC</span>
         </Link>
-        {navLinks.map((link) => (
-          <Link
-            key={link.path}
-            href={link.path}
-            className={cn(
-              'transition-colors hover:text-foreground',
-              checkActivePath(link.path)
-                ? 'text-foreground'
-                : 'text-muted-foreground',
-            )}
-          >
-            {link.name}
-          </Link>
-        ))}
-      </nav>
+        <Navigation />
+      </div>
       <Sheet>
         <SheetTrigger asChild>
           <Button variant="outline" size="icon" className="shrink-0 md:hidden">
